[src] outsource cross section calculations, new Physics_Model class.
