+++ CMakeLists.txt
@@ -25,11 +25,13 @@ add_library(thgem SHARED
   include/thread_pool.hpp
   include/fields.hh 
   include/geomodel.hh
+  include/pmodel.hh
   include/transport.hh 
   include/electrode.hh 
   src/thread_pool.cpp
   src/fields.cpp 
   src/geomodel.cpp 
+  src/pmodel.cpp 
   src/transport.cpp 
   src/electrode.cpp )
 target_link_libraries(thgem ${ROOT_LIBRARIES})
